:recycle: Move spring auto configuration to 'autoconfigure' package
