+++ middleware/sisyphus-amqp/src/main/kotlin/com/bybutter/sisyphus/middleware/amqp/autoconfigure/SisyphusAmqpAutoConfiguration.kt
@@ -1,11 +1,11 @@
-package com.bybutter.sisyphus.middleware.amqp
+package com.bybutter.sisyphus.middleware.amqp.autoconfigure
 
+import com.bybutter.sisyphus.middleware.amqp.AmqpTemplateFactory
+import com.bybutter.sisyphus.middleware.amqp.DefaultAmqpTemplateFactory
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.ComponentScan
-import org.springframework.context.annotation.Configuration
 
-@Configuration
 @ComponentScan(basePackageClasses = [SisyphusAmqpAutoConfiguration::class])
 class SisyphusAmqpAutoConfiguration {
     @Bean
