+++ middleware/sisyphus-amqp/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-com.bybutter.sisyphus.middleware.amqp.SisyphusAmqpAutoConfiguration
+com.bybutter.sisyphus.middleware.amqp.autoconfigure.SisyphusAmqpAutoConfiguration
