+++ middleware/sisyphus-grpc-client/src/main/kotlin/com/bybutter/sisyphus/middleware/grpc/ClientRepository.kt
@@ -1,8 +1,6 @@
 package com.bybutter.sisyphus.middleware.grpc
 
-import com.bybutter.sisyphus.rpc.CallOptionsInterceptor
-import com.bybutter.sisyphus.rpc.SisyphusStub
-import com.bybutter.sisyphus.spi.Ordered
+import com.bybutter.sisyphus.middleware.grpc.autoconfigure.ClientRegistrar
 import io.grpc.CallOptions
 import io.grpc.Channel
 import io.grpc.ClientInterceptor
@@ -13,7 +11,10 @@ import org.springframework.beans.factory.config.ConfigurableListableBeanFactory
 import org.springframework.beans.factory.support.AbstractBeanDefinition
 import org.springframework.core.env.Environment
 
-interface ClientRepository : Ordered {
+interface ClientRepository {
+
+    var order: Int
+
     fun listClientBeanDefinition(beanFactory: ConfigurableListableBeanFactory, environment: Environment): List<AbstractBeanDefinition>
 
     fun getStubFromService(service: Class<*>): Class<*> {
@@ -27,30 +28,39 @@ interface ClientRepository : Ordered {
                 ?: throw IllegalStateException("Grpc service must have nested class named 'Client'.")
     }
 
-    fun interceptStub(stub: AbstractStub<*>, builderInterceptors: Iterable<ClientBuilderInterceptor>, interceptors: Iterable<ClientInterceptor>): AbstractStub<*> {
+    fun processStub(stub: AbstractStub<*>, beanFactory: ConfigurableListableBeanFactory): AbstractStub<*> {
         var result = stub
 
-        for (interceptor in builderInterceptors) {
-            result = interceptor.intercept(result)
+        val builderInterceptors = beanFactory.getBeansOfType(ClientBuilderInterceptor::class.java)
+        for ((_, builderInterceptor) in builderInterceptors) {
+            result = builderInterceptor.intercept(result)
         }
 
-        return result.withInterceptors(*interceptors.toList().toTypedArray())
+        val interceptors = beanFactory.getBeansOfType(ClientInterceptor::class.java)
+        return result.withInterceptors(*interceptors.values.toTypedArray())
     }
 
-    fun createGrpcChannel(target: String, beanFactory: ConfigurableListableBeanFactory): Channel {
-        return ManagedChannelBuilder.forTarget(target).usePlaintext().userAgent("Generated by Sisyphus").build().apply {
+    fun createGrpcChannel(property: GrpcChannelProperty, beanFactory: ConfigurableListableBeanFactory): Channel {
+        return ManagedChannelBuilder.forTarget(property.target).usePlaintext().userAgent("Generated by Sisyphus").build().apply {
             val channelLifecycleManager = beanFactory.getBean(ClientRegistrar.QUALIFIER_AUTO_CONFIGURED_GRPC_CHANNEL_LIFECYCLE) as ManagedChannelLifecycle
             channelLifecycleManager.registerManagedChannel(this)
         }
     }
 
-    fun createGrpcClient(target: Class<*>, channel: Channel, optionsInterceptors: Iterable<CallOptionsInterceptor>, callOptions: CallOptions): AbstractStub<*> {
-        if (SisyphusStub::class.java.isAssignableFrom(target)) {
-            return target.getDeclaredConstructor(Channel::class.java, Iterable::class.java, CallOptions::class.java)
-                .newInstance(channel, optionsInterceptors, callOptions) as AbstractStub<*>
-        } else {
-            return target.getDeclaredConstructor(Channel::class.java, CallOptions::class.java)
-                .newInstance(channel, callOptions) as AbstractStub<*>
+    fun createGrpcChannel(name: String, port: Int, beanFactory: ConfigurableListableBeanFactory): Channel {
+        return ManagedChannelBuilder.forAddress(name, port).usePlaintext().userAgent("Generated by Sisyphus").build().apply {
+            val channelLifecycleManager = beanFactory.getBean(ClientRegistrar.QUALIFIER_AUTO_CONFIGURED_GRPC_CHANNEL_LIFECYCLE) as ManagedChannelLifecycle
+            channelLifecycleManager.registerManagedChannel(this)
         }
+    }
+
+    fun createGrpcClient(target: Class<*>, channel: Channel, property: GrpcChannelProperty): AbstractStub<*> {
+        return target.getDeclaredConstructor(Channel::class.java, CallOptions::class.java)
+                .newInstance(channel, property.options) as AbstractStub<*>
+    }
+
+    fun createGrpcClient(target: Class<*>, channel: Channel): AbstractStub<*> {
+        return target.getDeclaredConstructor(Channel::class.java, CallOptions::class.java)
+                .newInstance(channel, CallOptions.DEFAULT) as AbstractStub<*>
     }
 }
