+++ middleware/sisyphus-grpc-client/src/main/kotlin/com/bybutter/sisyphus/middleware/grpc/autoconfigure/SisyphusGrpcClientAutoConfiguration.kt
@@ -1,8 +1,6 @@
-package com.bybutter.sisyphus.middleware.grpc
+package com.bybutter.sisyphus.middleware.grpc.autoconfigure
 
 import org.springframework.context.annotation.ComponentScan
-import org.springframework.context.annotation.Configuration
 
-@Configuration
 @ComponentScan(basePackageClasses = [SisyphusGrpcClientAutoConfiguration::class])
 class SisyphusGrpcClientAutoConfiguration
