+++ middleware/sisyphus-grpc-client/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-com.bybutter.sisyphus.middleware.grpc.SisyphusGrpcClientAutoConfiguration
+com.bybutter.sisyphus.middleware.grpc.autoconfigure.SisyphusGrpcClientAutoConfiguration
