+++ middleware/sisyphus-redis/src/main/kotlin/com/bybutter/sisyphus/middleware/redis/autoconfigure/SisyphusRedisAutoConfiguration.kt
@@ -1,11 +1,11 @@
-package com.bybutter.sisyphus.middleware.redis
+package com.bybutter.sisyphus.middleware.redis.autoconfigure
 
+import com.bybutter.sisyphus.middleware.redis.DefaultRedisClientFactory
+import com.bybutter.sisyphus.middleware.redis.RedisClientFactory
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.ComponentScan
-import org.springframework.context.annotation.Configuration
 
-@Configuration
 @ComponentScan(basePackageClasses = [SisyphusRedisAutoConfiguration::class])
 class SisyphusRedisAutoConfiguration {
     @Bean
