+++ middleware/sisyphus-redis/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-com.bybutter.sisyphus.middleware.redis.SisyphusRedisAutoConfiguration
+com.bybutter.sisyphus.middleware.redis.autoconfigure.SisyphusRedisAutoConfiguration
