+++ starter/sisyphus-grpc-server-starter/src/main/kotlin/com/bybutter/sisyphus/starter/grpc/SisyphusGrpcServerAutoConfiguration.kt
@@ -1,11 +1,10 @@
 package com.bybutter.sisyphus.starter.grpc
 
-import com.bybutter.sisyphus.middleware.grpc.SisyphusGrpcClientAutoConfiguration
+import com.bybutter.sisyphus.middleware.grpc.autoconfigure.SisyphusGrpcClientAutoConfiguration
+import org.springframework.boot.autoconfigure.AutoConfigureAfter
 import org.springframework.boot.autoconfigure.AutoConfigureBefore
 import org.springframework.context.annotation.ComponentScan
-import org.springframework.context.annotation.Configuration
 
-@Configuration
 @AutoConfigureBefore(SisyphusGrpcClientAutoConfiguration::class)
 @ComponentScan(basePackageClasses = [SisyphusGrpcServerAutoConfiguration::class])
 class SisyphusGrpcServerAutoConfiguration
