+++ starter/sisyphus-protobuf-type-server-starter/src/main/kotlin/com/bybutter/sisyphus/starter/protobuf/type/SisyphusProtobufTypeServerAutoConfiguration.kt
@@ -1,11 +1,9 @@
 package com.bybutter.sisyphus.starter.protobuf.type
 
 import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
 import org.springframework.web.reactive.function.server.RouterFunction
 import org.springframework.web.reactive.function.server.ServerResponse
 
-@Configuration
 class SisyphusProtobufTypeServerAutoConfiguration {
     @Bean
     fun protobufTypeReflectionFunction(): RouterFunction<ServerResponse> {
